requirements.txt
python-dotenv
python-telegram-bot
pymongo
dnspython
